requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websocket-client>=1.7.0
python-dotenv>=1.0.0
streamlit>=1.33.0
requests>=2.31.0
fredapi >= 0.5.2
pandas>=2.2.0
plotly>=5.18.0
yfinance>=0.2.37
aiohttp>=3.9.0
websockets>=12.0
streamlit-autorefresh>=0.0.1
duckduckgo-search>=6.0.0
google-genai>=1.0.0
kaleido>=0.2.1
Pillow>=9.0.0
